requirements.txt
streamlit  
pandas  
numpy  
scikit-learn  
requests  
joblib  
openai-whisper